chore(node-config-provider): add preferred config file to load
